Allow external state callback to be specified via build flag...
... and allow main() to be called by external module.

Useful for UI implementations that simply layer on top of ui-console
(rather than reimplement it).  In such cases, ui-console can be built as
a library rather than top-level executable.  The higher-level UI can
then "launch" the ui-console implementation in its own process/thread
by calling main directly (rather than spawning a separate ui-console
process).  This is not only a general convenience, but also allows the
higher-level UI to provide a state callback to synchronize its display.

This commit facilitates downstream UI development, but is not likely to
be of interest to general users who build ui-console from makefile.
Therefore, the CALLBACK_* build flags are omitted from the makefile help
text.